refactor: update project structure
