+++ src/app.module.ts
@@ -4,8 +4,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { dataSourceOptions } from './data-source';
 import { GroupsModule } from './groups/groups.module';
 import { APP_FILTER, APP_INTERCEPTOR } from '@nestjs/core';
-import { LogInterceptor } from './log.interceptor';
-import { AllExceptionsFilter } from './utils/exceptions.filter';
+import { LogInterceptor } from './utils/interceptors/log.interceptor';
+import { AllExceptionsFilter } from './utils/filters/exceptions.filter';
 import { AuthModule } from './auth/auth.module';
 
 @Module({
