+++ src/utils/constants/entities.ts
